# Improve performance of memoized methods with arguments by ~50%