--- a/lib/memery.rb
+++ b/lib/memery.rb
@@ -80,7 +80,7 @@ module Memery
           end
 
           cache_store = (@_memery_memoized_values ||= {})
-          cache_key = original_arity.zero? ? method_key : [method_key, *args]
+          cache_key = original_arity.zero? ? method_key : [method_key, *args].hash
           cache = cache_store[cache_key]
 
           return cache.result if cache&.fresh?(ttl)