@deril 2024-09-12
> Based on the information in this twitter thread

You meant ex twitter, aka X